+++ src/br/com/projetoperiodo/model/instituto/disciplina/dao/JPADisciplinaDao.java
@@ -59,7 +59,7 @@ public class JPADisciplinaDao implements DisciplinaDao{
 	}
 
 	@Override
-	public Disciplina buscar(int primaryK) {
+	public Disciplina buscar(long primaryK) {
 		EntityManager entityManager =  JPAUtil.getInstance().getEntityManagerFactory().createEntityManager();
 		EntityTransaction entityTransaction = entityManager.getTransaction();
 		entityTransaction.begin();
